search feature

+++ src/main/java/com/backend/elearning/domain/order/OrderDetailPostDto.java
@@ -1,5 +1,5 @@
 package com.backend.elearning.domain.order;
 
 public record OrderDetailPostDto (Long courseId,
-                                  double price) {
+                                  Long price) {
 }
